initial file split and rename
